[Bug 1143941] Factor Questions and Localization to be two configurations of one UI.
