+++ kitsune/community/static/js/community-l10n.browserify.js
@@ -1,11 +1,25 @@
 /* jshint esnext: true */
-/* globals React:false */
-import L10nContributors from './L10nContributors.jsx';
 import CommunityController from './CommunityController.js';
+import {UserChip, RelativeTime} from './contributors-common.jsx';
 
 window.onpopstate = function() {
     refresh();
 }
 
-var controller = new CommunityController('l10n', L10nContributors, document.querySelector('#main-content'));
+var controller = new CommunityController({
+    area: 'l10n',
+    target: document.querySelector('#main-content'),
+    title: 'Top Contributors - Knowledge Base',
+    columns: [
+        {key: 'rank', title: 'Rank'},
+        {key: 'user', title: 'Name', transform: (u) => <UserChip {...u}/>},
+        {key: 'revision_count', title: 'Revisions'},
+        {key: 'review_count', title: 'Reviews'},
+        {
+            key: 'last_contribution_date',
+            title: 'Last Activity',
+            transform: (timestamp) => <RelativeTime timestamp={timestamp} future={false}/>
+        },
+    ],
+});
 controller.render();
