+++ kitsune/community/static/js/community-questions.browserify.js
@@ -1,11 +1,27 @@
 /* jshint esnext: true */
-/* globals React:false */
-import QuestionsContributors from './QuestionsContributors.jsx';
 import CommunityController from './CommunityController.js';
+import {UserChip, RelativeTime} from './contributors-common.jsx';
 
 window.onpopstate = function() {
     refresh();
 }
 
-var controller = new CommunityController('questions', QuestionsContributors, document.querySelector('#main-content'));
+var controller = new CommunityController({
+    area: 'questions',
+    target: document.querySelector('#main-content'),
+    title: 'Top Contributors - Questions',
+    columns: [
+        {key: 'rank', title: 'Rank'},
+        {key: 'user', title: 'Name', transform: (u) => <UserChip {...u}/>},
+        {key: 'answer_count', title: 'Answers'},
+        {key: 'solution_count', title: 'Solutions'},
+        {key: 'helpful_vote_count', title: 'Helpful Votes'},
+        {
+            key: 'last_contribution_date',
+            title: 'Last Activity',
+            transform: (timestamp) => <RelativeTime timestamp={timestamp} future={false}/>
+        },
+    ],
+})
+
 controller.render();
