+++ kitsune/community/static/less/community-new.less
@@ -36,6 +36,12 @@ table.top-contributors {
     [data-column="select"] {
         text-align: center;
     }
+
+    [data-column="actions"] {
+        padding-right: 10px;
+        text-align: center;
+        width: 0; // As small as possible.
+    }
 }
 
 .user-chip {
